+++ examples/pytorch/gat/main.py
@@ -12,7 +12,7 @@ from deepgnn.pytorch.common.dataset import TorchDeepGNNDataset
 from deepgnn.pytorch.modeling import BaseModel
 from deepgnn.pytorch.training import run_dist
 from deepgnn.graph_engine import FileNodeSampler, GraphEngineBackend
-from model import GAT, GATQueryParameter
+from model import GAT, GATQueryParameter  # type: ignore
 
 
 # fmt: off