+++ src/python/deepgnn/graph_engine/utils.py
@@ -1,14 +1,14 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 """Graph engine configuration."""
-
+import argparse
 from deepgnn.arg_types import str2bool
 from deepgnn.graph_engine.backends.options import GraphType
 from deepgnn.graph_engine.snark.converter.options import DataConverterType
 from deepgnn.graph_engine.graph_dataset import BackendType
 
 
-def define_param_graph_engine(parser):
+def define_param_graph_engine(parser: argparse.ArgumentParser):
     """Define command line arguments for graph engine."""
     group = parser.add_argument_group("Graph Client Parameters")
     group.add_argument(